requirements.txt
groq
apify-client
requests
python-dotenv
rich